Move staking utils to lib
